pyproject.toml
[project]
name = "todo-app-backend-25-02-20"
version = "0.1.0"
description = ""
authors = [
    {name = "elena7993",email = "elenalee7993@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
